# fastapi_server/openapi (2).json
{"openapi":"3.1.0","info":{"title":"Banco REST API - Práctica 3","description":"API REST para la gestión de operaciones bancarias, transicionada desde Java RMI.","contact":{"name":"José Lis Martín Vera","email":"joseluis.martin@ejemplo.com"},"version":"1.0.0"},"paths":{"/sesiones":{"post":{"tags":["Sesiones"],"summary":"Paso 1: Identificación de usuario","description":"Inicia el proceso de login identificando al usuario.\nDevuelve un `session_id` que debe usarse en las siguientes peticiones.","operationId":"identificar_sesiones_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentificacionRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sesiones/{session_id}/auth":{"put":{"tags":["Sesiones"],"summary":"Paso 2: Autenticación con contraseña","description":"Actualiza el estado de la sesión validando la contraseña.\nTransición: sesión no autenticada → sesión autenticada.\nSe usa PUT porque modifica el estado de un recurso existente (/sesiones/{session_id}).","operationId":"autenticar_sesiones__session_id__auth_put","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AutenticacionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sesiones/{session_id}/lotes":{"post":{"tags":["Operaciones"],"summary":"Paso 3: Envío de lote de acciones","description":"Envía una cadena de instrucciones para ser analizadas por el servidor.\nEl lote queda pendiente de confirmación hasta que el cliente responda.","operationId":"procesar_lote_sesiones__session_id__lotes_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sesiones/{session_id}/lotes/confirmacion":{"post":{"tags":["Operaciones"],"summary":"Paso 4: Confirmación de operaciones","description":"Confirma o cancela el lote de operaciones pendiente para la sesión indicada.","operationId":"confirmar_sesiones__session_id__lotes_confirmacion_post","parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmacionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AutenticacionRequest":{"properties":{"password":{"type":"string","title":"Password","description":"Contraseña del cliente"}},"type":"object","required":["password"],"title":"AutenticacionRequest"},"BatchRequest":{"properties":{"instrucciones":{"type":"string","title":"Instrucciones","description":"Cadena de acciones separadas por comas (ej: '1, 2 100')"}},"type":"object","required":["instrucciones"],"title":"BatchRequest"},"ConfirmacionRequest":{"properties":{"decision":{"type":"string","title":"Decision","description":"Decisión de confirmar ('si' o 'no')"}},"type":"object","required":["decision"],"title":"ConfirmacionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IdentificacionRequest":{"properties":{"username":{"type":"string","title":"Username","description":"Nombre de usuario del cliente"}},"type":"object","required":["username"],"title":"IdentificacionRequest"},"ResponseMessage":{"properties":{"mensaje":{"type":"string","title":"Mensaje"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"},"resumen":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resumen"},"saldo_final":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Saldo Final"}},"type":"object","required":["mensaje"],"title":"ResponseMessage"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}